+++ graph/graph-application/src/main/kotlin/eu/tib/orkg/prototype/statements/api/resource-queries.kt
@@ -40,7 +40,12 @@ interface RetrieveResourceUseCase {
         label: SearchString,
         pageable: Pageable
     ): Page<Resource>
+
+    /** Find any resource by DOI that has one of the publishable classes. */
     fun findByDOI(doi: String): Optional<Resource>
+
+    /** Find a paper resource by DOI. */
+    fun findPaperByDOI(doi: String): Optional<Resource>
     fun findById(id: ThingId): Optional<Resource>
     fun findByTitle(title: String): Optional<Resource>
     fun findAllComparisonsByObservatoryId(id: ObservatoryId, pageable: Pageable): Page<Resource>