+++ testing/spring/src/testFixtures/kotlin/eu/tib/orkg/prototype/testing/AsciidocExtensions.kt
@@ -1,0 +1,3 @@
+package eu.tib.orkg.prototype.testing
+
+fun Iterable<String>.toAsciidoc(): String = this.joinToString(separator = ", ", prefix = "`", postfix = "`")